fix cycling dependencies
