+++ idunn/api/instant_answer.py
@@ -20,7 +20,7 @@ from idunn.utils.regions import get_region_lonlat
 from idunn.utils.result_filter import ResultFilter
 from idunn.instant_answer import normalize
 from .constants import PoiSource
-from .utils import Verbosity
+from ..utils.verbosity import Verbosity
 
 logger = logging.getLogger(__name__)
 result_filter = ResultFilter()
