+++ idunn/places/base.py
@@ -4,7 +4,7 @@ from geopy import Point
 from pytz import timezone, UTC
 from typing import Optional, Union
 
-from idunn.api.utils import Verbosity, build_blocks
+from ..utils.verbosity import build_blocks, Verbosity
 from idunn.datasources.wiki_es import wiki_es
 from idunn.utils import maps_urls, tz
 from .place import Place, PlaceMeta
